textEmbedding/requirements.txt
grpcio
protobuf
grpcio-reflection
grpcio-status
torch==2.3.1
torchvision==0.18.1
numpy==1.26.1
scipy
Pillow
huggingface_hub
einops
safetensors
sentence-transformers